+++ src/OFS/PropertyManager.py
@@ -21,8 +21,6 @@ from AccessControl.Permissions import manage_properties
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from Acquisition import aq_base
 from ExtensionClass import Base
-from App.special_dtml import DTMLFile
-from App.Dialogs import MessageDialog
 from zExceptions import BadRequest
 from zope.interface import implements
 from ZPublisher.Converters import type_converters
@@ -77,22 +75,6 @@ class PropertyManager(Base):
 
     Entries in the _properties structure which do not have a 'mode' key
     are assumed to have the mode 'wd' (writeable and deleteable).
-
-    To fully support property management, including the system-provided
-    tabs and user interfaces for working with properties, an object which
-    inherits from PropertyManager should include the following entry in
-    its manage_options structure::
-
-      {'label':'Properties', 'action':'manage_propertiesForm',}
-
-    to ensure that a 'Properties' tab is displayed in its management
-    interface. Objects that inherit from PropertyManager should also
-    include the following entry in its __ac_permissions__ structure::
-
-      ('Manage properties', ('manage_addProperty',
-                             'manage_editProperties',
-                             'manage_delProperties',
-                             'manage_changeProperties',)),
     """
 
     implements(IPropertyManager)
@@ -102,15 +84,7 @@ class PropertyManager(Base):
     security.setPermissionDefault(access_contents_information,
                                   ('Anonymous', 'Manager'))
 
-    manage_options = (
-        {'label': 'Properties', 'action': 'manage_propertiesForm'},
-    )
-
-    security.declareProtected(manage_properties, 'manage_propertiesForm')
-    manage_propertiesForm = DTMLFile(
-        'dtml/properties', globals(), property_extensible_schema__=1)
-    security.declareProtected(manage_properties, 'manage_propertyTypeForm')
-    manage_propertyTypeForm = DTMLFile('dtml/propertyType', globals())
+    manage_options = tuple()
 
     title = ''
     _properties = (
@@ -280,8 +254,6 @@ class PropertyManager(Base):
         if type in type_converters:
             value = type_converters[type](value)
         self._setProperty(id.strip(), value, type)
-        if REQUEST is not None:
-            return self.manage_propertiesForm(self, REQUEST)
 
     security.declareProtected(manage_properties, 'manage_editProperties')
     def manage_editProperties(self, REQUEST):
@@ -300,10 +272,6 @@ class PropertyManager(Base):
                 else:
                     value = REQUEST.form.get(name, '')
                 self._updateProperty(name, value)
-        if REQUEST:
-            message = "Saved changes."
-            return self.manage_propertiesForm(self, REQUEST,
-                                              manage_tabs_message=message)
 
     security.declareProtected(manage_properties, 'manage_changeProperties')
     def manage_changeProperties(self, REQUEST=None, **kw):
@@ -327,10 +295,6 @@ class PropertyManager(Base):
                 if 'w' not in propdict[name].get('mode', 'wd'):
                     raise BadRequest('%s cannot be changed' % escape(name))
                 self._updateProperty(name, value)
-        if REQUEST:
-            message = "Saved changes."
-            return self.manage_propertiesForm(self, REQUEST,
-                                              manage_tabs_message=message)
 
     security.declareProtected(manage_properties, 'manage_changePropertyTypes')
     def manage_changePropertyTypes(self, old_ids, props, REQUEST=None):
@@ -348,8 +312,6 @@ class PropertyManager(Base):
             return
         for prop in props:
             self._setProperty(prop.new_id, prop.new_value, prop.new_type)
-        if REQUEST is not None:
-            return self.manage_propertiesForm(self, REQUEST)
 
     security.declareProtected(manage_properties, 'manage_delProperties')
     def manage_delProperties(self, ids=None, REQUEST=None):
@@ -360,10 +322,7 @@ class PropertyManager(Base):
                 ids = None
             ids = REQUEST.get('_ids', ids)
         if ids is None:
-            return MessageDialog(
-                title='No property specified',
-                message='No properties were specified!',
-                action='./manage_propertiesForm')
+            raise BadRequest('No propery specified.')
         propdict = self.propdict()
         nd = self._reserved_names
         for id in ids:
@@ -371,14 +330,7 @@ class PropertyManager(Base):
                 raise BadRequest(
                     'The property <em>%s</em> does not exist' % escape(id))
             if ('d' not in propdict[id].get('mode', 'wd')) or (id in nd):
-                return MessageDialog(
-                    title='Cannot delete %s' % id,
-                    message='The property <em>%s</em> '
-                            'cannot be deleted.' % escape(id),
-                    action='manage_propertiesForm')
+                raise BadRequest('Cannot delete %s' % id)
             self._delProperty(id)
-
-        if REQUEST is not None:
-            return self.manage_propertiesForm(self, REQUEST)
 
 InitializeClass(PropertyManager)